+++ samples/effect/bin/start-server.ts
@@ -1,33 +1,13 @@
+import p from "../src/server"
 import { effect as T, managed as M, exit as E } from "@matechs/effect"
 import * as KOA from "@matechs/koa"
 // import { Do } from "fp-ts-contrib/lib/Do"
 import { pipe } from "fp-ts/lib/pipeable"
-import { sequenceT } from "fp-ts/lib/Apply"
-
-const routeA = KOA.route(
-  "get",
-  "/",
-  T.pure(
-    KOA.routeResponse(200, {
-      message: "OK",
-    }),
-  ),
-)
-
-const mainR = sequenceT(T.effect)(routeA)
-//const subR = pipe(sequenceT(T.effect)(routeC, routeD), KOA.withSubRouter("/sub"))
 
 const port = 3535
 
 const program = pipe(
-  sequenceT(T.effect)(
-    mainR,
-    //subR,
-    KOA.middleware((ctx, next) => {
-      ctx.set("X-Request-Id", "my-id")
-      return next()
-    }),
-  ),
+  p,
   // keep process waiting
   T.chainTap(() => T.never),
   M.provideS(KOA.managedKoa(port)),
