+++ samples/effect/nodemon.json
@@ -1,12 +1,5 @@
 {
-  "watch": [
-    ".",
-    "src/**/*",
-    "../../packages/**/*"
-  ],
-  "ignore": [
-    "data/**/*",
-    "router-schema.json"
-  ],
+  "watch": [".", "bin/**/*", "src/**/*", "../../packages/**/*"],
+  "ignore": ["data/**/*", "router-schema.json"],
   "ext": "ts, js, json"
 }
